refactor: separate CapTP communications from trap iterator driver
