+++ packages/captp/lib/captp.js
@@ -15,13 +15,11 @@ import { isPromise } from '@agoric/promise-kit';
 import { assert, details as X } from '@agoric/assert';
 
 import { makeTrap, nearTrapImpl } from './trap.js';
+import { makeTrapGuest, makeTrapHost } from './trap-driver.js';
 
 import './types.js';
 
 export { E };
-
-// Something to make our reply sequences more interesting.
-const TRAP_FIRST_SEQUENCE_NUMBER = 23;
 
 /**
  * @template T
@@ -43,13 +41,6 @@ const TRAP_FIRST_SEQUENCE_NUMBER = 23;
  * (host) to serve objects marked with makeTrapHandler to synchronous clients
  * (guests)
  */
-
-/**
- * @param {any} maybeThenable
- * @returns {boolean}
- */
-const isThenable = maybeThenable =>
-  maybeThenable && typeof maybeThenable.then === 'function';
 
 /**
  * Create a CapTP connection.
@@ -118,6 +109,9 @@ export function makeCapTP(ourId, rawSend, bootstrapObj = undefined, opts = {}) {
     },
   );
 
+  const trapHost = giveTrapReply && makeTrapHost();
+  const trapGuest = takeTrapReply && makeTrapGuest(unserialize);
+
   /** @type {WeakMap<any, string>} */
   const valToSlot = new WeakMap(); // exports looked up by val
   const slotToVal = new Map(); // reverse
@@ -136,8 +130,6 @@ export function makeCapTP(ourId, rawSend, bootstrapObj = undefined, opts = {}) {
   const questions = new Map(); // chosen by us
   /** @type {Map<number, any>} */
   const answers = new Map(); // chosen by our peer
-  /** @type {Map<number, (data: any, nonce: number) => void>} */
-  const answerToTrapNextBuffer = new Map();
   /** @type {Map<number, any>} */
   const imports = new Map(); // chosen by our peer
 
@@ -380,45 +372,18 @@ export function makeCapTP(ourId, rawSend, bootstrapObj = undefined, opts = {}) {
           assert.typeof(
             giveTrapReply,
             'function',
-            X`CapTP internal error; cannot answer Trap(x) without opts.giveTrapReply`,
+            X`CapTP cannot answer Trap(x) without opts.giveTrapReply`,
           );
+          assert(trapHost, X`CatTP internal error; trapHost is not defined`);
+          sendReturn = async (isReject, value) => {
+            const serialized = serialize(harden(value));
+            const it = giveTrapReply(isReject, serialized);
+            await trapHost.sendTrapReply(questionID, it);
+          };
         } catch (e) {
           sendReturn(true, e);
           throw e;
         }
-        sendReturn = async (isReject, value) => {
-          // sendTrap may return an iterator.
-          const serialized = serialize(harden(value));
-          const it = giveTrapReply(isReject, serialized);
-
-          // Allow the caller to iterate via `CTP_TRAP_NEXT_BUFFER` messages.
-          let nextSeq = TRAP_FIRST_SEQUENCE_NUMBER;
-          let pendingP;
-          /**
-           * @param {number} seq
-           * @param {any} data
-           */
-          const nextInSequence = async (seq, data) => {
-            // Prevent parallel requests for a given sequence number.
-            await pendingP;
-            assert.equal(
-              seq,
-              nextSeq,
-              X`Invalid CTP_TRAP_NEXT_BUFFER; seq ${seq} is not ${nextSeq}`,
-            );
-            nextSeq += 1;
-            pendingP = it.next(data);
-            const status = await pendingP;
-            if (status.done) {
-              // No more "nexts" are required.
-              answerToTrapNextBuffer.delete(questionID);
-            }
-          };
-          answerToTrapNextBuffer.set(questionID, nextInSequence);
-
-          // Prime the pump with the first iteration.
-          await nextInSequence(nextSeq, undefined);
-        };
       }
 
       // If `args` is supplied, we're applying a method or function... otherwise this is
@@ -442,13 +407,9 @@ export function makeCapTP(ourId, rawSend, bootstrapObj = undefined, opts = {}) {
         .catch(rej => sendReturn(true, rej))
         .catch(rej => quietReject(rej, false));
     },
-    // Send the next part of a "trap" call's result.
-    async CTP_TRAP_NEXT_BUFFER(obj) {
-      const { questionID, data, seq } = obj;
-      const nextInSequence = answerToTrapNextBuffer.get(questionID);
-      assert(nextInSequence);
-      return nextInSequence(seq, data);
-    },
+    // Have the host serve the next buffer request.
+    CTP_TRAP_NEXT_BUFFER:
+      trapHost && (obj => trapHost.trapNextBuffer(obj.questionID, obj.trapBuf)),
     // Answer to one of our questions.
     async CTP_RETURN(obj) {
       const { result, exception, answerID } = obj;
@@ -604,27 +565,13 @@ export function makeCapTP(ourId, rawSend, bootstrapObj = undefined, opts = {}) {
         }
       }
 
-      // Set up the trap call with its identifying information.
-      const it = takeTrapReply(implMethod, slot, implArgs);
-      const nextTrap = prep => {
-        const status = it.next(prep);
-        assert(
-          !isThenable(status),
-          X`takeTrapReply must be a fully-synchronous Generator but it.next() returned a Thenable ${status}`,
-        );
-        return status;
-      };
+      assert(trapGuest, X`Trap(${target}) internal error; no trapGuest`);
 
-      // Prepare the reply, in case there needs to be something initialized
-      // before the call is ready.
-      let firstTime = true;
-      let status = nextTrap(firstTime);
-      let seq = TRAP_FIRST_SEQUENCE_NUMBER;
-      while (!status.done) {
-        if (firstTime) {
-          // Mark the send as a "trap", but handle it asynchronously on the
-          // other side.
-          firstTime = false;
+      // Set up the trap call with its identifying information.
+      const takeIt = takeTrapReply(implMethod, slot, implArgs);
+      return trapGuest.doTrap(
+        takeIt,
+        () =>
           send({
             type: 'CTP_CALL',
             epoch,
@@ -632,34 +579,15 @@ export function makeCapTP(ourId, rawSend, bootstrapObj = undefined, opts = {}) {
             questionID,
             target: slot,
             method,
-          });
-        } else {
-          // We weren't done fetching the entire result in the first iteration,
-          // so thread through the iterator values into the "next" messages to
-          // retrieve the entire serialized data.
-          seq += 1;
+          }),
+        trapBuf =>
           send({
             type: 'CTP_TRAP_NEXT_BUFFER',
             epoch,
             questionID,
-            seq,
-            data: status.value,
-          });
-        }
-        status = nextTrap(firstTime);
-      }
-
-      // We've finished claiming the return value.
-      const [isReject, serialized] = status.value;
-      const value = unserialize(serialized);
-      assert(
-        !isThenable(value),
-        X`Trap() reply cannot be a Thenable; have ${value}`,
+            trapBuf,
+          }),
       );
-      if (isReject) {
-        throw value;
-      }
-      return value;
     };
 
     /** @type {TrapImpl} */
